apps/rooms/api_endpoints/List/serializers.py
from rest_framework import serializers

from apps.rooms.models import Room


class RoomListSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = ("id", "name", "type", "capacity")
