requirements/base.txt
asgiref==3.6.0
Django==4.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-environ==0.4.54
pytz==2023.3
sqlparse==0.4.3
psycopg2-binary==2.9.5
drf-yasg==1.21.5
django-parler==2.3
Pillow==9.5.0
pre-commit==3.3.2
django-filter==23.2